restructure for esbuild & to update pagedjs.
